managed-pandera-msp-main/sandboxcreationandcleanup-main/projects-msp/msp-cleanup-admin/src_code/requirements.txt
google-cloud-resource-manager==1.5.1
google-cloud-billing==1.3.0
sendgrid